web/trulens-dashboard/requirements.txt
trulens-eval==0.25.1
psycopg2-binary
streamlit==1.32.0